backend/requirements.txt
fastapi
uvicorn
sqlmodel
h3
shapely
scikit-learn
numpy
requests
langchain
langchain-groq
duckdb
python-dotenv
groq
pandas
sqlparse
# PostGIS/PostgreSQL dependencies
psycopg2-binary
asyncpg
sqlalchemy[asyncio]
geoalchemy2
